requirements.txt
flask==3.0.0
pymongo==4.6.1
nltk==3.8.1
pyspellchecker==0.7.2
langdetect==1.0.9
pandas==2.1.4
imblearn==0.0
bs4==0.0.1
textblob==0.17.1
gensim==4.3.2
sentence-transformers==2.1.0
flask_cors==4.0.0
flask-bcrypt==1.0.1
PyJWT==2.8.0
